Decimal version

+++ Program.cs
@@ -19,19 +19,19 @@ namespace MyApp
         static void Main(string[] args)
         {
             int quantity = 20; // integer scalar representing the quantity of goods
-            double pricePerItem = 4.95; // assign the price per item to $4.95
-            const float salesTaxPercentage = 0.05F; // floating point type constant
+            decimal pricePerItem = 4.95M; // assign the price per item to $4.95
+            const decimal salesTaxPercentage = 0.05M; // floating point type constant
 
             // sub total is just quantity multiplied by the price per item
-            double subTotal = quantity * pricePerItem; // note the type coercion
+            decimal subTotal = quantity * pricePerItem; // note the type coercion
                                                        // promote quantity
 
             // calculate sales tax - not the cast to explicitly
             // change the data type of salesTaxPercentage to a double
             // if we did not do that, C# would implicitly convert anyways
-            double salesTax = subTotal * (double)salesTaxPercentage;
+            decimal salesTax = subTotal * salesTaxPercentage;
 
-            double total = subTotal + salesTax;
+            decimal total = subTotal + salesTax;
 
             Console.WriteLine("The total is " + total);
